requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
lxml==4.9.3
python-dotenv==1.0.0
schedule==1.2.0
webdriver-manager==4.0.1
